Fixed missing Storybook QueryClientProvider
