--- .storybook/preview.ts
@@ -1,17 +1,0 @@
-import type { Preview } from "@storybook/react";
-
-import "../src/styles/main.css";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  tags: ["autodocs"],
-};
-
-export default preview;
