+++ .storybook/preview.tsx
@@ -1,0 +1,30 @@
+import type { Preview } from "@storybook/react";
+import React from "react";
+import { QueryClient, QueryClientProvider } from "react-query";
+
+import "../src/styles/main.css";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  tags: ["autodocs"],
+  decorators: [
+    (Story) => {
+      const queryClient = new QueryClient();
+
+      return (
+        <QueryClientProvider client={queryClient}>
+          <Story />
+        </QueryClientProvider>
+      );
+    },
+  ],
+};
+
+export default preview;
